heartwall fixes

+++ openmp/heartwall/AVI/avilib.c
@@ -83,7 +83,7 @@ static size_t avi_write (int fd, char *buf, size_t len)
 
    while (r < len) {
       n = write (fd, buf + r, len - r);
-      if (n < 0)
+      if ((int) n < 0)
          return n;
       
       r += n;
@@ -157,7 +157,7 @@ static int avi_add_chunk(avi_t *AVI, unsigned char *tag, unsigned char *data, in
    length = PAD_EVEN(length);
 
    if( avi_write(AVI->fdes,(char *)c,8) != 8 ||
-       avi_write(AVI->fdes,(char *)data,length) != length )
+       (int) avi_write(AVI->fdes,(char *)data,length) != length )
    {
       lseek(AVI->fdes,AVI->pos,SEEK_SET);
       AVI_errno = AVI_ERR_WRITE;
@@ -543,7 +543,7 @@ int avi_update_header(avi_t *AVI)
    
    //11/14/01 added id string 
 
-   if(njunk > strlen(id_str)+8) {
+   if(njunk > (int) strlen(id_str)+8) {
      sprintf(id_str, "%s-%s", PACKAGE, VERSION);
      memcpy(AVI_header+nhb, id_str, strlen(id_str));
    }
@@ -1177,7 +1177,7 @@ int avi_parse_input_file(avi_t *AVI, int getIndex)
 	    
 	    header_offset = lseek(AVI->fdes,0,SEEK_CUR);
 				 
-            if( avi_read(AVI->fdes,(char *)hdrl_data,n) != n ) ERR_EXIT(AVI_ERR_READ)
+            if( (long) avi_read(AVI->fdes,(char *)hdrl_data,n) != n ) ERR_EXIT(AVI_ERR_READ)
          }
          else if(strncasecmp(data,"movi",4) == 0)
          {
@@ -1195,7 +1195,7 @@ int avi_parse_input_file(avi_t *AVI, int getIndex)
          AVI->n_idx = AVI->max_idx = n/16;
          AVI->idx = (unsigned  char((*)[16]) ) malloc(n);
          if(AVI->idx==0) ERR_EXIT(AVI_ERR_NO_MEM)
-         if(avi_read(AVI->fdes, (char *) AVI->idx, n) != n ) ERR_EXIT(AVI_ERR_READ)
+         if((long) avi_read(AVI->fdes, (char *) AVI->idx, n) != n ) ERR_EXIT(AVI_ERR_READ)
       }
       else
          lseek(AVI->fdes,n,SEEK_CUR);
@@ -1350,7 +1350,7 @@ int avi_parse_input_file(avi_t *AVI, int getIndex)
       lseek(AVI->fdes,pos,SEEK_SET);
       if(avi_read(AVI->fdes,data,8)!=8) ERR_EXIT(AVI_ERR_READ)
       if( strncasecmp((char *)data,(char *)AVI->idx[i],4)==0 && 
-      str2ulong((unsigned char *)data+4)==len )
+      (long) str2ulong((unsigned char *)data+4)==len )
       {
          idx_type = 1; /* Index from start of file */
       }
@@ -1358,7 +1358,7 @@ int avi_parse_input_file(avi_t *AVI, int getIndex)
       {
          lseek(AVI->fdes,pos+AVI->movi_start-4,SEEK_SET);
          if(avi_read(AVI->fdes,data,8)!=8) ERR_EXIT(AVI_ERR_READ)
-         if( strncasecmp((char *)data,(char *)AVI->idx[i],4)==0 && str2ulong((unsigned char *)data+4)==len )
+         if( (long) strncasecmp((char *)data,(char *)AVI->idx[i],4)==0 && (long) str2ulong((unsigned char *)data+4)==len )
          {
             idx_type = 2; /* Index from start of movi list */
          }
@@ -1620,7 +1620,7 @@ long AVI_read_frame(avi_t *AVI, char *vidbuf, int *keyframe)
    n = AVI->video_index[AVI->video_pos].len;
    *keyframe = (AVI->video_index[AVI->video_pos].key==0x10) ? 1:0;
    lseek(AVI->fdes, AVI->video_index[AVI->video_pos].pos, SEEK_SET);
-   if (avi_read(AVI->fdes,vidbuf,n) != n)
+   if ( (long) avi_read(AVI->fdes,vidbuf,n) != n)
    {
       AVI_errno = AVI_ERR_READ;
       return -1;
@@ -1646,7 +1646,7 @@ int AVI_set_audio_position(avi_t *AVI, long byte)
    while(n0<n1-1)
    {
       n = (n0+n1)/2;
-      if(AVI->track[AVI->aptr].audio_index[n].tot>byte)
+      if((long)AVI->track[AVI->aptr].audio_index[n].tot > byte)
          n1 = n;
       else
          n0 = n;
@@ -1683,7 +1683,7 @@ long AVI_read_audio(avi_t *AVI, char *audbuf, long bytes)
          todo = left;
       pos = AVI->track[AVI->aptr].audio_index[AVI->track[AVI->aptr].audio_posc].pos + AVI->track[AVI->aptr].audio_posb;
       lseek(AVI->fdes, pos, SEEK_SET);
-      if (avi_read(AVI->fdes,audbuf+nr,todo) != todo)
+      if ((long)avi_read(AVI->fdes,audbuf+nr,todo) != todo)
       {
          AVI_errno = AVI_ERR_READ;
          return -1;
@@ -1744,7 +1744,7 @@ int AVI_read_data(avi_t *AVI, char *vidbuf, long max_vidbuf,
             lseek(AVI->fdes,n,SEEK_CUR);
             return -1;
          }
-         if(avi_read(AVI->fdes,vidbuf,n) != n ) return 0;
+         if((long)avi_read(AVI->fdes,vidbuf,n) != n ) return 0;
          return 1;
       }
       else if(strncasecmp(data,AVI->track[AVI->aptr].audio_tag,4) == 0)
@@ -1755,7 +1755,7 @@ int AVI_read_data(avi_t *AVI, char *vidbuf, long max_vidbuf,
             lseek(AVI->fdes,n,SEEK_CUR);
             return -2;
          }
-         if(avi_read(AVI->fdes,audbuf,n) != n ) return 0;
+         if((long)avi_read(AVI->fdes,audbuf,n) != n ) return 0;
          return 2;
          break;
       }
